Cover all guide code blocks as verified sample regions
Every code block in first-endpoint.md is now a named region in
FirstEndpointSamples.kt, including:
- imports: the full import block with canonical package paths
- hello-server-test: runBlocking wrapper showing correct test idiom
- greet-server-test: complete greet endpoint test (was prose-only before)
- echo-server-test: typed endpoint test returning structured output

The prose notes that bind is a ServerBuilder member (no import), that
the implementation lambda has a ServerRuntime context, and that runBlocking
is needed because the test {} action is not suspend but .test() calls are.

FirstEndpointTest now delegates to the sample functions directly rather than
duplicating assertions, so there is one source of truth.

The guide directory is declared as a test task input in build.gradle.kts so
editing any .md file invalidates the Gradle cache and re-runs DriftCheckTest
automatically without --rerun-tasks.
