+++ docs-guide/src/samples/kotlin/com/lightningkite/lightningserver/guide/samples/FirstEndpointSamples.kt
@@ -1,8 +1,12 @@
 package com.lightningkite.lightningserver.guide.samples
 
+// region imports
 import com.lightningkite.lightningserver.auth.noAuth
 import com.lightningkite.lightningserver.definition.builder.ServerBuilder
-import com.lightningkite.lightningserver.http.*
+import com.lightningkite.lightningserver.http.HttpHandler
+import com.lightningkite.lightningserver.http.HttpResponse
+import com.lightningkite.lightningserver.http.get
+import com.lightningkite.lightningserver.http.post
 import com.lightningkite.lightningserver.pathing.arg1
 import com.lightningkite.lightningserver.plainText
 import com.lightningkite.lightningserver.runtime.test.test
@@ -10,6 +14,7 @@ import com.lightningkite.lightningserver.typed.ApiHttpHandler
 import com.lightningkite.lightningserver.typed.test
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.Serializable
+// endregion imports
 
 // region hello-server
 object HelloServer : ServerBuilder() {
@@ -21,6 +26,15 @@ object HelloServer : ServerBuilder() {
 }
 // endregion hello-server
 
+// region hello-server-test
+fun helloServerTest() = runBlocking {
+    HelloServer.test(settings = {}) {
+        val response = HelloServer.root.test()
+        check(response.body?.text() == "Hello, Lightning Server!")
+    }
+}
+// endregion hello-server-test
+
 // region greet-server
 object GreetServer : ServerBuilder() {
 
@@ -31,6 +45,15 @@ object GreetServer : ServerBuilder() {
     }
 }
 // endregion greet-server
+
+// region greet-server-test
+fun greetServerTest() = runBlocking {
+    GreetServer.test(settings = {}) {
+        val response = GreetServer.greet.test("World")
+        check(response.body?.text() == "Hello, World!")
+    }
+}
+// endregion greet-server-test
 
 // region echo-types
 @Serializable
@@ -58,12 +81,21 @@ object EchoServer : ServerBuilder() {
 }
 // endregion echo-server
 
-// Top-level function referenced by @sample in ApiHttpHandler KDoc.
-// Shows the idiomatic way to define and test a typed, unauthenticated endpoint.
-fun echoServerSample() = runBlocking {
+// region echo-server-test
+fun echoServerTest() = runBlocking {
     EchoServer.test(settings = {}) {
         // ApiHttpHandler.test() accepts null auth for noAuth endpoints and
         // returns the typed output directly — no JSON manipulation needed.
+        val result = EchoServer.echo.test(null, EchoRequest("ping"))
+        check(result.echo == "ping")
+        check(result.length == 4)
+    }
+}
+// endregion echo-server-test
+
+// Top-level function referenced by @sample in ApiHttpHandler KDoc.
+fun echoServerSample() = runBlocking {
+    EchoServer.test(settings = {}) {
         val result = EchoServer.echo.test(null, EchoRequest("ping"))
         check(result.echo == "ping")
         check(result.length == 4)
